+++ src/patterns.js
@@ -6,14 +6,15 @@
  * only want to use selected patterns you will need to pull in the patterns
  * directly in your RequireJS configuration.
  */
+
 define([
     "jquery",
     "pat-registry",
+    "moment",
     "modernizr",
     "prefixfree",
 //    "push-kit",
     "pat-ajax",
-
     "pat-autofocus",
     "pat-autoscale",
     "pat-autosubmit",
@@ -24,13 +25,14 @@ define([
     "pat-carousel",
     "pat-carousel-legacy",
     "pat-checklist",
-    "pat-chosen",
+//    "pat-chosen",
     "pat-clone",
     "pat-collapsible",
     "pat-colour-picker",
     "pat-date-picker",
     "pat-datetime-picker",
     "pat-depends",
+    "pat-display-time",
     "pat-equaliser",
     "pat-expandable",
     "pat-focus",
@@ -65,6 +67,31 @@ define([
     "pat-validation",
     "pat-zoom"
 ], function($, registry) {
+
+    import("moment-locale-de");  // jshint ignore:line
+
+    // import("moment-locale-bg");
+    // import("moment-locale-hr");
+    // import("moment-locale-cs");
+    // import("moment-locale-da");
+    // import("moment-locale-nl");
+    // import("moment-locale-es");
+    // import("moment-locale-fi");
+    // import("moment-locale-fr");
+    // import("moment-locale-el");
+    // import("moment-locale-hu");
+    // import("moment-locale-it");
+    // import("moment-locale-lt");
+    // import("moment-locale-lv");
+    // import("moment-locale-mt");
+    // import("moment-locale-pl");
+    // import("moment-locale-pt");
+    // import("moment-locale-ro");
+    // import("moment-locale-sl");
+    // import("moment-locale-sk");
+    // import("moment-locale-es");
+    // import("moment-locale-sv");
+
     // Since we are in a non-AMD env, register a few useful utilites
     var window = require("window");
     window.jQuery = $;