+++ webpack/karma.config.js
@@ -3,6 +3,8 @@
 // See more on https://www.npmjs.com/package/webpack-karma-jasmine
 const path = require('path');
 var webpackOptions = require('./base.config.js');
+webpackOptions['mode'] = 'development';
+process.env.CHROME_BIN = require('puppeteer').executablePath();
 
 module.exports = function(config) {
     config.set({
@@ -43,15 +45,20 @@ module.exports = function(config) {
             //use webpack to support require() in test-suits .js files
             //use babel-loader from webpack to compile es2015 features in .js files
             //add webpack as preprocessor
-            //'test_requires.js': ['webpack'],
             'tests/specs/*/*.js': ['webpack', 'sourcemap' /*, 'coverage'*/ ],
             'src/pat/*/tests.js': ['webpack', 'sourcemap' /*, 'coverage'*/ ],
             // list here the files that you want transpiled by babel
             'src/pat/masonry/tests.js': ['webpack', 'sourcemap', 'babel' /*, 'coverage'*/ ],
+            'src/pat/calendar/tests.js': ['webpack', 'sourcemap', 'babel' /*, 'coverage'*/ ],
         },
         babelPreprocessor: {
             options: {
-                presets: ['env'],
+                plugins: ['syntax-dynamic-import'],
+                presets: [["@babel/env", {
+                      "targets": {
+                        "browsers": ["last 2 versions", "ie >= 11"]
+                      }
+                    }]],
                 sourceMap: 'inline'
             },
             filename: function(file) {
@@ -106,7 +113,7 @@ module.exports = function(config) {
 
         // start these browsers
         // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
-        browsers: ['Chrome' /*,'PhantomJS','Firefox','Edge','ChromeCanary','Opera','IE','Safari'*/ ],
+        browsers: ['MyCustomLauncher' /*,'PhantomJS','Firefox','Edge','ChromeCanary','Opera','IE','Safari'*/ ],
 
 
         // Continuous Integration mode
@@ -134,9 +141,15 @@ module.exports = function(config) {
         },
 
         customLaunchers: {
-            'PhantomJS_debug': {
-                base: 'PhantomJS',
-                debug: true
+            'MyCustomLauncher': {
+                base: 'ChromeHeadless',
+                debug: true,
+                options: {
+                    viewportSize: {
+                      width: 1440,
+                      height: 1200
+                    }
+                }
             }
         }
     })