+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 4.0.0 - unreleased
+
+- Runs now on jQuery 3
+- Integrated pat-display-time from https://github.com/ploneintranet/pat-display-time
+- fix minimum input length default so that you can display select results already on click
+
 ## 3.0.0a5 - unreleased
 
 Features
@@ -13,6 +19,7 @@ Features
 Fixes
 ~~~~~
 
+- pat-date-picker, pat-datetime-picker: Support the `first-day` parameter (#647)
 - pat-notification: fix how the close button is rendered (#639)
 - pat-modal: remove an handler after the modal is closed (allows for injection inside modals, see #550)
 - Enable babel transpiler